Updates and function changes
There is no longer a need to choose a host when selecting a cluster that has DRS enabled; the host will be chosen automatically based on lowest node usage.

There was an issue with datastore selections. If selecting a datastore that was part of a datastore cluster but did not have automatic Sdrs enabled, VMs would fail to be built. This has been rectified.

There are now individual (editable) drop downs for both vCPU sockets and vCPU cores, so you can specify them independently.

The folder selection panel in the Add VM dialog has been changed. There is now a button that you need to click in order to populate the folder frame. This is to help mitigate some of the performance issue in the GUI until I can really dig down into some of the performance improvements.

Textboxes have been enabled to allow editing of NIC card labels, so you label network adapters at build time (i.e. FrontEnd, BackEnd, etc.). These textboxes can be found on the AddVM dialog, the DuplicateVM dialog and the Add Additional NICs dialog.